refactor(1682 #6): collapse reporters/ package → single module
#6 structure cleanup (#1682): `reporters/` is over-packaging — an EMPTY
`__init__.py` wrapping one real module (`console.py`, 207 L), only 1 importer.
Collapse the thin package: move console.py's body into `__init__.py` so
`from reyn.reporters import X` is the surface (the __init__ holds the content).

- git rm the empty `__init__.py`; git-mv `console.py` → `__init__.py`
  (rename-tracked, preserves console.py's history).
- the lone importer (cli/logger_factory.py) used the DEEP path
  `from reyn.reporters.console import ConsoleLogger` → rewritten to the package
  surface `from reyn.reporters import ConsoleLogger` (the deep `.console` path is
  gone). 1 one-line edit.

Only reporters/ collapsed — the other thin dirs (budget/dispatch/etc.) interact
with the #5 grouping decision (owner-held), left untouched per the dispatch.

Verified: `from reyn.reporters import ConsoleLogger` + logger_factory import OK;
ruff src tests clean. No dedicated reporters tests exist (thin module); full-suite
CI covers regression. behavior-preserving.

Refs #1682 (#6).
